### 2025-02-20 #7807 @editorialbot
[PRE REVIEW]: Sugar: A Python framework for bioinformatics
**Submitting author:** @trichter (<a href="http://orcid.org/0000-0002-8378-559X">Tom Eulenfeld</a>)
**Repository:** https://github.com/rnajena/sugar
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.0
