### 2025-02-20 #7807 @trichter
A suitable reviewer might be padix-key.

I added a LICENSE file to the repository.
