### 2025-02-20 863976b
ftsviewer: automatically convert label to str
