### 2025-02-20 42af038
select: use eq if no operator given
