### 2025-02-24 1445a3f
rf='all' is now the default in find_orfs
