### 2025-02-25 a2ef82d
add test creating another fts plot
