### 2025-02-25 4ea56bf
extend and harden find_orfs
