+++ src/sugar/core/cane.py
@@ -154,11 +154,18 @@ class BioMatch(object):
         return getattr(self._match, attr)
     def __repr__(self):
         return f'<sugar.BioMatch object; match={self.group()}; span={self.span()}; rf={self.rf}; seqid={self.seqid}>'
-    def span(self):
-        start, stop = self._match.span()
+    def start(self):
         if self.rf is not None and self.rf < 0:
-            start, stop = self.lenseq - stop, self.lenseq - start
-        return start, stop
+            return self.lenseq - self._match.end()
+        else:
+            return self._match.start()
+    def end(self):
+        if self.rf is not None and self.rf < 0:
+            return self.lenseq - self._match.start()
+        else:
+            return self._match.end()
+    def span(self):
+        return self.start(), self.end()
 
 
 class BioMatchList(collections.UserList):
@@ -175,6 +182,17 @@ class BioMatchList(collections.UserList):
         :return: Nested dict structure
         """
         return _groupby(self, keys)
+
+    def select(self, **kw):
+        """
+        Select matches
+
+        :param keys: Tuple of meta keys or functions to use for grouping.
+            Can also be a single string or a callable.
+            By default the method groups by seqid only.
+        :return: Nested dict structure
+        """
+        return _select(self, attr=None, **kw)
 
     @property
     def d(self):
@@ -220,7 +238,10 @@ def match(seq, sub, *, rf='fwd',
             Defaults to False.
 
     Returns:
-        match (`BioMatch` or `BioMatchList` of matches or None)
+        match (`BioMatch` or `BioMatchList` of matches or None),
+        the list will be sorted by match position,
+        matches on the forward strand first,
+        then matches on the backward strand.
     """
     from bisect import bisect
     import re
@@ -280,7 +301,6 @@ def match(seq, sub, *, rf='fwd',
                         matches.append(m)
                     else:
                         return m
-
     if matchall:
         return BioMatchList(matches)
 
@@ -292,7 +312,7 @@ class ORFList(FeatureList):
     pass
 
 
-def _inds2orf(i1, i2, rf, lensec, ftype='ORF', seqid=None):
+def _inds2orf(i1, i2, rf, lensec, ftype='ORF', seqid=None, has_start=None, has_stop=None):
     """
     Create ORF feature from indices
     """
@@ -307,11 +327,20 @@ def _inds2orf(i1, i2, rf, lensec, ftype='ORF', seqid=None):
     ft.seqid = seqid
     ft.loc.strand = strand
     ft.meta.rf = rf
+    ft.meta.has_start = has_start
+    ft.meta.has_stop = has_stop
     return ft
 
 
+def _get_from_list(l, i, default=None):
+    try:
+        return l[i]
+    except IndexError:
+        return default
+
+
 def find_orfs(seq, rf='all', start='start', stop='stop', need_start='always', need_stop=True,
-              allow_within=False, gap='-', minlen=0, ftype='ORF'):
+              nested='other', gap='-', len_ge=0, ftype='ORF'):
     """
     Find open reading frames (ORFs)
 
@@ -320,17 +349,20 @@ def find_orfs(seq, rf='all', start='start', stop='stop', need_start='always', ne
         `~match`. Default is ``'all'``.
     :param start: regular expression defining the start codons, defaults to ATG/AUG
     :param stop: regular expression defining the stop codons, defaults to stop codons
-        in default translation table
+        in the default translation table
     :param need_start: One of ``('always', 'once', 'never')``.
-        Always: Ech ORF starts with a start codon.
-        Once: Only the first ORF in each RF starts with a start codon.
+        Always: Each ORF starts with a start codon.
+        Once: Only the first ORF on the forward and backward strand starts with a start codon.
[…]
