+++ src/sugar/tests/test_core_cane.py
@@ -73,15 +73,153 @@ def test_find_orfs():
     orfs2 = seqs[0].find_orfs(rf='all')
     assert len(orfs2) > len(orfs)
 
-    orfs3 = seqs[0].find_orfs(rf='fwd', allow_within=True)
-    assert len(orfs3) > len(orfs)
+    orfsb = seqs[0].find_orfs(rf='fwd', nested='all')
+    assert len(orfsb) > len(orfs)
 
-    orfs4 = seqs[0].find_orfs(rf='all', allow_within=True)
+    orfsc = seqs[0].find_orfs(rf='fwd', need_stop=False)
+    assert len(orfsc) == len(orfs) + 2
+
+    orfs4 = seqs[0].find_orfs(rf='all', nested='all')
     assert len(orfs4) > len(orfs2)
 
     orfs = seqs.find_orfs()
     for id_ in seqs.ids:
         assert seqs.d[id_][orfs.groupby('seqid')[id_].sort(len)[-1]] == seqs.d[id_]['cds']
+
+
+def test_find_orfs_extended():
+    teststr = 'CCC_CAT_GCT_GAC_CTA_CCC_CAT_CAT_GTG_ACC_CCC_CTG_ACC_CAT_GCC_CTG_ACC_ATG_CCC_C'
+    seq = BioSeq(teststr.replace('_', ''))
+    assert (seq.find_orfs(rf='fwd') + seq.find_orfs(rf='bwd')) == seq.find_orfs(rf='all')
+    orfs1 = seq.find_orfs()
+    orfs2 = seq.find_orfs(need_start='never')
+    orfs3 = seq.find_orfs(need_start='once')
+    orfs1b = seq.find_orfs(nested='all')
+    orfs2b = seq.find_orfs(need_start='never', nested='all')
+    orfs3b = seq.find_orfs(need_start='once', nested='all')
+    orfs1c = seq.find_orfs(nested='no')
+    orfs2c = seq.find_orfs(need_start='never', nested='no')
+    orfs3c = seq.find_orfs(need_start='once', nested='no')
+    assert len(orfs1c) < len(orfs1) < len(orfs1b)
+    assert len(orfs1c) == 3
+    assert len(orfs1) == 4
+    assert len(orfs1b) == 7
+    assert orfs2 == orfs2b
+    assert len(orfs2) == 6
+    assert len(orfs2c) == 2
+    assert orfs3 == orfs3b
+    assert len(orfs3) == 6
+    assert len(orfs3c) == 3
+    assert orfs1c == orfs1c.copy().sort(lambda ft: [0, 1, 2, -1, -2, -3].index(ft.meta.rf))
+    orfs3d = seq.copy().rc().find_orfs(need_start='once', nested='all')
+    for orf in orfs3d:
+        orf.meta.rf = - 1 - orf.meta.rf
+    assert orfs3d.copy().rc(seqlen=len(seq)).sort() == orfs3b.copy().sort()
+    teststr3 = teststr[:30] + '-----' + teststr[30:]
+    seq = BioSeq(teststr3.replace('_', ''))
+    assert (seq.find_orfs(rf='fwd') + seq.find_orfs(rf='bwd')) == seq.find_orfs(rf='all')
+    assert len(orfs1) == len(seq.find_orfs())
+    assert len(orfs2) == len(seq.find_orfs(need_start='never'))
+    assert len(orfs3) == len(seq.find_orfs(need_start='once'))
+    assert len(orfs1b) == len(seq.find_orfs(nested='all'))
+    assert len(orfs2b) == len(seq.find_orfs(need_start='never', nested='all'))
+    assert len(orfs3b) == len(seq.find_orfs(need_start='once', nested='all'))
[…]
