+++ docs/src/tutorial_imaging.rst
@@ -78,9 +78,9 @@ Another example plotting the CDS features of an Ebolavirus RNA:
 .. image:: ../_static/fts3.png
     :width: 60%
 
-For comparison we plot all open reading frames longer than 500 nucleotides:
+For comparison we plot all open reading frames with a minimum length of 500 nucleotides:
 
->>> orfs = seq.find_orfs().select(len_gt=500)
+>>> orfs = seq.find_orfs(len_ge=500)
 >>> orfs.plot_ftsviewer(colorby='rf', label=None, seqlen=len(seq), figsize=(6, 2.5), show=True)
 
 .. image:: ../_static/fts4.png
