### 2025-02-26 bebc164
add test for remove_overlapping and remove_nested
