### 2025-02-26 2a25a7c
add color argument in tutorial
