### 2025-02-27 b136156
raise clearer error message + test
