### 2025-02-27 07dc174
seqlen fix + test
