### 2025-02-27 32da0a7
another tiny fix + test
