### 2025-02-27 112aa95
decrease runtime in find_orfs x5, just by not using pop(0)
