### 2025-03-03 22ebab3
quote only necessary characters in GFF3 writer
