+++ src/sugar/tests/test_io_gff.py
@@ -1,7 +1,7 @@
 # (C) 2024, Tom Eulenfeld, MIT license
 
 from sugar import read, read_fts
-from sugar.tests.util import _clean_seqs, _clean_fts, tempfilename
+from sugar.tests.util import _clean_fts, tempfilename
 
 
 def test_gff_fts():
@@ -21,6 +21,18 @@ def test_gtf():
         fts.write(fname)
         fts2 = _clean_fts(read_fts(fname))
     assert fts == fts2
+
+
+def test_gff_quote():
+    fts = read_fts('!data/fts_example.gff')
+    fts[0].seqid = '%,;&'
+    fts[0].name = ';-), :-)'
+    with tempfilename(suffix='.gff') as fname:
+        fts.write(fname)
+        fts2 = read_fts(fname)
+    assert fts2[0].seqid == fts[0].seqid
+    assert fts2[0].name == fts[0].name
+    assert ':' in fts.tofmtstr('gff')
 
 
     # TODO test write and load again
