### 2025-03-04 75e381e
add note about color vs colorby
