### 2025-03-07 ccc0c67
allow to set date for logo
