### 2025-03-18 9def888
add option to select Feature by strand
