+++ CONTRIBUTING.md
@@ -1,0 +1,30 @@
+## How to contribute
+
+Thank you for considering contributing to this project!
+
+There are several ways to contribute. We list some of them here in order of increasing difficulty.
+
+### Discuss new features, ask questions
+
+Not much to say here. Just use the issue tracker or contact the developers by mail.
+
+
+### Report bugs
+
+You can report bugs using the issue tracker on GitHub. Please make sure that 1. you are running the latest version of the package and 2. that the test suite runs smoothly with your installation (`sugar test` command).
+
+Optionally, you can try to check if the bug is already fixed in the development version by installing it (see the documentation on how to install the dev version).
+
+When reporting bugs, it is best to include a minimal reproducible example and, if available, an exception traceback. Please also format code blocks or tracebacks properly by enclosing them with triple back ticks.
+
+### Improve the documentation
+
+Minor errors in the documentation can be fixed directly in the GitHub editor by clicking the edit button in the top right corner. If asked, please submit the changes as a pull request.
+
+Changes to the doc strings inside the source code cannot be fixed this way. Please see the next paragraph for that.
+
+### Improve the code
+
+Glad that you made it this far ;).
+To improve the code base, you will need to fork the GitHub repository, commit your improvements to a new branch, and send a pull request to the main repository. Please try to add a test for every new feature or bugfix.
+Please also read the documentation on how to write new IO plugins, how to build the documentation locally, etc.
