### 2025-03-19 a8960b3
mention contributing file in readme
