### 2025-03-19 #7807 @trichter
@Kevin-Mattheus-Moerman Dear Kevin, thank you for these first hints. I added more of context to the project description and also added a contributing file with more detailed guidelines.
