### 2025-03-24 d6ba548
add outdir fixture and command line option for writing files needed in the documentation to a given directory
