+++ src/sugar/tests/conftest.py
@@ -1,11 +1,16 @@
 from contextlib import chdir
 from copy import deepcopy
+from pathlib import Path
 import pytest
 
 
 def pytest_addoption(parser):
     parser.addoption('--web', action='store_true',
                  default=False, help='enable web tests')
+    msg = ('output directory for files generated by the test suite, '
+           'e.g. images used in the documentation')
+    parser.addoption('--outdir', help=msg)
+
 
 def pytest_collection_modifyitems(config, items):
     if not config.getoption('--web'):
@@ -49,3 +54,12 @@ def tmp_path_cd(tmp_path):
     """
     with chdir(tmp_path):
         yield tmp_path
+
+
+@pytest.fixture()
+def outdir(request, tmp_path):
+    """
+    A fixture which changes the dir to a temporary directory
+    """
+    outdir = request.config.getoption('--outdir')
+    return tmp_path if outdir is None else Path(outdir)
