### 2025-03-24 c0503f3
add tmpfname pytest fixture to return the name of a temp file, either use that fixture or the tmpfilename context manager to read,write named temporary files within tests
