### 2025-04-09 e5ec859
make python version requirement explicit
