### 2025-04-09 981251f
update license field, remove license classifier
