+++ pyproject.toml
@@ -5,10 +5,9 @@ authors = [
     {name = "Tom Eulenfeld"},
 ]
 description = "A Python framework for bioinformatics"
-license = {text = "MIT License"}
+license = "MIT"
 classifiers = [
     "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
 keywords = [
