### 2025-04-10 0cbd3cb
add lite version of dna_fts_viewer as dep
