### 2025-04-10 8b7d13e
update ci: use lite version
