+++ pyproject.toml
@@ -7,8 +7,11 @@ authors = [
 description = "A Python framework for bioinformatics"
 license = "MIT"
 classifiers = [
-    "Programming Language :: Python :: 3",
+    "Intended Audience :: Science/Research",
+    "Intended Audience :: Education",
     "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Topic :: Scientific/Engineering :: Bio-Informatics"
 ]
 keywords = [
     "bioinformatics", "RNA", "DNA", "sequence",
