### 2025-04-16 #7807 @olexandr-konovalov
Hello @lskatz and @erinyoung - would one or both of you be able to review this submission for JOSS, please?

If not, could you possibly recommend someone whom we may approach regarding this? Thanks!
