### 2025-04-24 25d9cc5
blast: old blast files with sstrand N/A and start larger than stop can be read now
