### 2025-04-24 331270c
gff3: allow lines to end with ;
