### 2025-04-29 #8122 @trichter
@tushardave26, thank you for the positive feedback and for the very quick review! Thank you also for the comments!

> However, I was unable to test the examples in the Sequence Features section because I did not have suitable example files available locally. To enhance the user experience and facilitate easier exploration of Sugar’s full capabilities, I strongly recommend including small, representative example files for each supported format directly within the repository.

Most of the functions could be tested using the example objects which are returned when `read()` and `read_fts()` are called without any argument. Does this work for you?

In fact, example files for several formats are already included in the repository [here](https://github.com/rnajena/sugar/tree/master/src/sugar/tests/data), but more for testing purposes. They can be read with the `!data` magic, e.g. `read('!data/example.fasta')` or `read_fts('!data/fts_example.gff')`, but this is not documented and only used in the tests.
I hesitate to advertize this functionality broadly, as it is more of a convenience for the developers of the package and also some files in this folder cover edgecases and are not representative.

> It would be beneficial to include a practical example or a short tutorial demonstrating how users can develop and integrate their own plugins.

This is documented in the [wiki](https://github.com/rnajena/sugar/wiki/How-to-create-a-new-IO-plugin). I have considered moving this developer-centric documentation into the main documentation.
Do you think this would make it easier to find? I would be interested in your opinion: Is this plugin documentation sufficient, i.e. is it useful for an advanced developer?

> While the documentation provides detailed examples, including a few key usage examples directly in the GitHub README would make the project more immediately accessible and easier for users to get started.

OK, thank you, I will work on that.

> Adding a Dockerfile would simplify environment setup and ensure reproducibility, allowing users to build and run Sugar quickly and consistently.

I am a bit reluctant to provide a dockerfile, as this dockerfile also needs to be maintained. This is a pure Python package so far, with no compilation required. But if you think that is definitely useful, I can provide one.
Currently, each release is published on pypi and additionally the corresponding repository is archived on zenodo.
