### 2025-04-29 #8122 @tushardave26
Thank you, @trichter, for your detailed and thoughtful response!

> Most of the functions could be tested using the example objects which are returned when `read()` and `read_fts()` are called without any argument. Does this work for you?

I tried it, and yes, it works very well to create example objects. Thank you for clarifying that!

> In fact, example files for several formats are already included in the repository [here](https://github.com/rnajena/sugar/tree/master/src/sugar/tests/data), but more for testing purposes. They can be read with the `!data` magic, e.g. `read('!data/example.fasta')` or `read_fts('!data/fts_example.gff')`, but this is not documented and only used in the tests.
> I hesitate to advertize this functionality broadly, as it is more of a convenience for the developers of the package and also some files in this folder cover edgecases and are not representative.

I also appreciate you highlighting that example files are already included. I now understand that the files contain a mix of regular examples and files designed for edge cases.

> This is documented in the [wiki](https://github.com/rnajena/sugar/wiki/How-to-create-a-new-IO-plugin). I have considered moving this developer-centric documentation into the main documentation.
> Do you think this would make it easier to find? I would be interested in your opinion: Is this plugin documentation sufficient, i.e. is it useful for an advanced developer?

Thank you for pointing me to the wiki! After reviewing it, moving this documentation into the primary project documentation would make it much easier to find. Integrating it into the primary docs (with perhaps a short "Extending Sugar" or "Plugin Development" section) would improve discoverability and make the package feel even more polished and welcoming for contributors.

> I am a bit reluctant to provide a dockerfile, as this dockerfile also needs to be maintained. This is a pure Python package so far, with no compilation required. But if you think that is definitely useful, I can provide one.
> Currently, each release is published on pypi and additionally the corresponding repository is archived on zenodo.

I completely understand your hesitation about the maintenance burden. However, I noticed that you are already using GitHub Actions for continuous integration. You can extend the existing CI workflow to build and publish a Docker image automatically. For example, you can trigger the action on each push to the `main` branch, build the Docker image, and then push it to a container registry (such as Docker Hub or GitHub Container Registry). This way, it would require minimal ongoing maintenance while offering users an additional convenient deployment option.
That said, I agree it’s a **“nice to have”** rather than a critical need. I just wanted to highlight that automation could make it relatively painless if you choose to add it.
