### 2025-05-07 f6ed193
add developer documentation (moved from wiki)
