### 2025-05-12 rnajena/sugar#2 @widdowquinn
SyntaxWarning/test skips/xfails when running `sugar test`
Part of https://github.com/openjournals/joss-reviews/issues/8122

In a clean conda environment built with Python 3.13 (macOS), I installed sugar from source using `pip install -U -e .` Running `sugar test` produced the warnings below:

```bash
% sugar test                                                           [10:29:10]
/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/web/_entrez.py:115: SyntaxWarning: invalid escape sequence '\*'
  :param \*\*kw: All other kwargs are passed to `fetch_seq()`
/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/web/_entrez.py:126: SyntaxWarning: invalid escape sequence '\*'
  :param \*\*kw: All other kwargs are passed to `fetch_seq()`
/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/web/_entrez.py:140: SyntaxWarning: invalid escape sequence '\*'
  :param \*\*kw: All other kwargs are passed to `fetch_basket()`
```

13 tests were skipped, and two xfailed.

```text
test_core_seq.py ...ssss.ss........................                                                             [ 43%]
test_imaging_alignment.py .s                                                                                    [ 50%]
test_imaging_fts.py ..s                                                                                         [ 52%]
test_index_fastaindex.py xx                                                                                     [ 54%]
test_io.py ............sss                                                                                      [ 64%]
test_web.py ss                                                                                                  [100%]

================================================== warnings summary ===================================================
test_core_fts.py::test_fts_toftsviewer
  /Users/lpritc/Library/Mobile Documents/com~apple~CloudDocs/Strathclyde_Work_iCloud/Refereeing/2025-05-12_JOSS/sugar/src/sugar/imaging/ftsviewer.py:37: SyntaxWarning: invalid escape sequence '\*'
    :param \*\*kw: Other kwargs are passed to

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
================================ 136 passed, 13 skipped, 2 xfailed, 1 warning in 8.50s ================================
```

Installation from PyPI using `pip install rnajena-sugar` in a clean Python 3.13 conda environment gives a slightly different set of skips/xfails, and no SyntaxWarnings.

```bash
% sugar test
test_core_fts.py ...............s..                                                                                                            [ 20%]
test_core_seq.py ...sssssss........................                                                                                            [ 43%]
test_imaging_alignment.py .s                                                                                                                   [ 50%]
test_imaging_fts.py sss                                                                                                                        [ 52%]
test_index_fastaindex.py xx                                                                                                                    [ 54%]
test_io.py ............sss                                                                                                                     [ 64%]
test_web.py ss                                                                                                                                 [100%]
```

While I might expect differences while the `main` branch is ahead of the PyPI release, I think it would be helpful for future users to have a reassuring note that some tests might be skipped or expected to fail, depending on circumstance.

The help message with `sugar test -h` does imply that the `--web` option is necessary for the `test_web.py` tests to run. With this option, the tests are not skipped. The `README.md` and documentation currently advises how to run tests, which is good, but I would like as a user for it to be explicit that skips/xfails are expected (and how to trigger those tests where desired), to avoid an incorrect assumption that there are problems with the codebase.
