### 2025-05-12 rnajena/sugar#3 @widdowquinn
Tutorial examples do not work as written.
Part of https://github.com/openjournals/joss-reviews/issues/8122

I was working through the tutorial at https://rnajena-sugar.readthedocs.io/en/latest/src/tutorial_seqs.html to try to understand the scope and utility of `sugar`. I would generally expect examples in a tutorial to run as written, and as intended, for the user. In examples such as:

```python
from sugar import read
seqs = read('example.fasta')
more_seqs = read('virus.stk', 'stockholm')
combined = seqs + more_seqs
combined.write('combined.fasta')
```

named files such as `virus.stk` are not found within the repository and are not clearly available to the reader. If a user does not have suitable files to hand, they cannot follow the tutorial as written. It may hinder uptake of the package if potential users cannot verify that it works as intended.

To support the user experience and enable them to complete the tutorial as written, I would suggest either to make these example files available (e.g. for download, or as package assets), or to modify the tutorial such that the user is directed to other publicly-available datasets for testing.
