### 2025-05-12 rnajena/sugar#4 @widdowquinn
Add documentation for use of `read()` with no argument.
Part of https://github.com/openjournals/joss-reviews/issues/8122

In the `README.md` for `sugar` an example is provided which uses the `read()` function with no arguments to generate an example dataset, rather than read data from file.

The inline documentation does not currently mention this capability. The use of `read()` without an argument to generate data appears to me to be unintuitive, and I would usually generally expect an error to be returned in this case, to guard against bugs where the sequence file has been accidentally omitted. As such I think it would be important to users that this behaviour is explicitly documented.

The current inline help reads to me that `fname` is a required argument, and the description string refers only to reading from file-like objects/files:

```text
Help on function read in module sugar._io.main:

read(fname, fmt=None, *, mode='r', encoding=None, **kw)
    Read a file or file-like object with sequences into `.BioBasket`

    :param fname: filename, can also be a glob expression,
        a web resource,
        an archive, gzipped file,
        or a file-like object (e.g. `~io.BytesIO`, `~io.StringIO`)
```
