### 2025-05-12 rnajena/sugar#3 @widdowquinn
A similar issue arises in https://rnajena-sugar.readthedocs.io/en/latest/src/tutorial_fts.html - the feature files used are not available, and I think it would be helpful to users if the tutorial could be followed as written.

I note that the Advanced Example at https://rnajena-sugar.readthedocs.io/en/latest/src/tutorial_seqs.html involves loading of an online resource as part of the tutorial (and it works as-written for me). This is the kind of tutorial/documentation I'd be looking for as a user.
