### 2025-05-12 rnajena/sugar#5 @widdowquinn
Add support for `join`, `order`, `complement` when parsing GenBank files?
While testing for https://github.com/openjournals/joss-reviews/issues/8122 my first check on a GenBank file using `read()` threw an error. It seems like location parsing does not yet accommodate all possibilities present in GenBank files. See, e.g. https://ftp.ncbi.nlm.nih.gov/genomes/all/GCA/040/267/505/GCA_040267505.1_UoG_LM41_1.0/ for an example.

```python
>>> from sugar import read
>>> genbank = read("GCA_040267505.1_UoG_LM41_1.0_genomic.gbff")
/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/_io/genbank.py:24: UserWarning: Parsing of genbank loc join, order, complement is untested
  warn('Parsing of genbank loc join, order, complement is untested')
Traceback (most recent call last):
  File "<python-input-6>", line 1, in <module>
    genbank = read("GCA_040267505.1_UoG_LM41_1.0_genomic.gbff")
  File "/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/_io/main.py", line 220, in new_reader
    return reader(fl, *args, **kw)
  File "/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/_io/main.py", line 316, in read
    seqs = list(getattr(module, funcname)(f, **kw))
  File "/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/_io/genbank.py", line 181, in iter_genbank
    ft = Feature(type=fttype, locs=_parse_locs(locs),
                                   ~~~~~~~~~~~^^^^^^
  File "/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/_io/genbank.py", line 26, in _parse_locs
    locs = [_parse_locs(subloc.strip())
            ~~~~~~~~~~~^^^^^^^^^^^^^^^^
  File "/Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/_io/genbank.py", line 28, in _parse_locs
    loc[loc.index('(')+1:loc.rindex(')')].split(',')]
                         ~~~~~~~~~~^^^^^
ValueError: substring not found
```

BioPython's `SeqIO` is currently able to parse these, and may be a useful reference for adding this capability.

```python
>>> from Bio import SeqIO
>>> records = list(SeqIO.parse("GCA_040267505.1_UoG_LM41_1.0_genomic.gbff", "genbank"))
>>> [len(_.features) for _ in records]
[16419, 339]
```
