### 2025-05-12 rnajena/sugar#6 @widdowquinn
Alignment plot output does not match tutorial.
As part of https://github.com/openjournals/joss-reviews/issues/8122 I followed the tutorial at https://rnajena-sugar.readthedocs.io/en/latest/src/tutorial_imaging.html.

The output image for `ali.pdf` is shown as:

![Example `ali.pdf` graphical output from the tutorial.](https://rnajena-sugar.readthedocs.io/en/latest/_images/ali2.png)

But the actual output as attached is quite different, and, I think, less useful.

```python
>>> seqs = read('https://osf.io/download/j2wyv')
>>> seqs.plot_alignment(show=True, color='gray', figsize=(10, 4))
<Axes: >
>>> seqs[:, 70:120].plot_alignment(fname='ali.pdf', color=None, figsize=(10,8), symbols=True, aspect=\
2, alpha=0.5)
```

[ali.pdf](https://github.com/user-attachments/files/20159824/ali.pdf)
