### 2025-05-12 rnajena/sugar#6 @widdowquinn
The next example in the same part of the tutorial has similar issues.

![Translated sequence example image from the tutorial](https://rnajena-sugar.readthedocs.io/en/latest/_images/ali3.png)

```python
>>> seqs2 = seqs[:5, :150].copy()
>>> seqs2.translate(complete=True).plot_alignment(show=True, color='flower', figsize=(10,8),  symbols\
=True,aspect=2, alpha=0.5, edgecolors='w')
2025-05-12 11:58:03.168 python[76054:46327308] The class 'NSSavePanel' overrides the method identifier.  This method is implemented by class 'NSWindow'
<Axes: >
```

![Image produced by code above](https://github.com/user-attachments/assets/6179c7fb-3194-4744-8690-b3240775e859)
