### 2025-05-12 rnajena/sugar#6 @widdowquinn
The feature viewer plots in the same tutorial render as expected.
