### 2025-05-12 rnajena/sugar#7 @widdowquinn
Preserve features when converting from BioPython record to BioBasket?
As part of https://github.com/openjournals/joss-reviews/issues/8122 I tried to work around #5 by reading GenBank records into BioPython, then converting using `sugar`. Unfortunately, this does not appear to preserve feature annotations:

```python
>>> from Bio import SeqIO
>>> records = list(SeqIO.parse("GCA_040267505.1_UoG_LM41_1.0_genomic.gbff", "genbank"))
>>> [len(_.features) for _ in records]
[16419, 339]
>>> from sugar import BioBasket
>>> seqs = BioBasket.frombiopython(records)
>>> print(seqs.fts[:10])
```

If it could be implemented, this might be a way to avoid problems when parsing complex GenBank locations using `sugar` - and it would be a neat feature to have, in general.
