### 2025-05-12 rnajena/sugar#8 @widdowquinn
How to parse GFF files that do not contain sequence directives?
As part of https://github.com/openjournals/joss-reviews/issues/8122 I attempted to read a standard `.gff` file that had the corresponding `.fna` file in the same directory using `read()`. This failed:

```python
>>> from sugar import read
>>> seqs = read("GCA_040267505.1_UoG_LM41_1.0_genomic.gff")
/Users/[REDATED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/core/seq.py:830: UserWarning: Features for seqids CP125626.1, CP125627.1 could not be attached to any sequence
  warn(f'Features for seqids {missing_ids} could not be '
```

Reading features with `read_fts` does work:

```python
>>> from sugar import read_fts
>>> fts = read_fts("GCA_040267505.1_UoG_LM41_1.0_genomic.gff")
>>> print(fts[:5])
region   0+ 7_784_629  name=ANONYMOUS;id=CP125626.1:1..7784629;seqid=CP125626...
  gene   0+       245  name=QMP25_00005;id=gene-QMP25_00005;seqid=CP125626.1;...
   CDS   0+       245  name=XBX14462.1;id=cds-XBX14462.1;seqid=CP125626.1;_gf...
  gene 354+       144  name=QMP25_00010;id=gene-QMP25_00010;seqid=CP125626.1;...
   CDS 354+       144  name=XBX14463.1;id=cds-XBX14463.1;seqid=CP125626.1;_gf...
```

but it is not obvious from the tutorial how to link data where the GFF and FASTA data are in distinct files. I assume that this would be possible to build as a plugin, accepting two filenames, but is there a way do link this data appropriately already, in `sugar`?
