### 2025-05-12 #8122 @widdowquinn
Hi @trichter - thank you for submitting `sugar` for review. This looks to be a very useful package, and I appreciate the effort it took to build. I was most excited to see the chaining style of data manipulation, built-in visualisation (especially for sequence alignments), and extendable plug-in framework. I am sure I will find a use for it in my own work, and when teaching students.

As you will see, I have raised some issues at https://github.com/rnajena/sugar/issues related to this review. They are not all blockers, but I thought I should note the issues I found while experimenting.

I have not yet checked two of the boxes relating to functionality. This is because of issues rnajena/sugar#4, rnajena/sugar#5, rnajena/sugar#7, and rnajena/sugar#8. For publication I don't think that it's critical the issues are solved to _deliver_ that function, but as a user I would expect to see clear flagging in the documentation that, for instance, complex GenBank locations may not be parsed correctly, that conversion between representations in different frameworks does not preserve data (ideally a statement of what is preserved would be helpful), and how to link GFF features to corresponding FASTA files (or a note that this cannot easily be done, so use the sequence directive). I think that these are sufficiently common tasks that, on finding that they do not work as expected and the behaviour is undocumented, users may move to other frameworks.

The other box currently unchecked relates to example usage. As a user I would expect that a tutorial would make it easy to obtain the data for me to work through examples and verify the package/tools. In some cases here, the tutorial data is accessible through weblinks, but in others it is not present. I think that the tutorial needs to have this data available to meet the spirit of those real-world examples required in the checklist.

Although I would consider that the functionality issues can be addressed by updates to and warnings in the documentation - I do think that providing the functionality would be of great help to potential users, and increase the reach of this package.

Congratulations on producing such a nice framework. I look forward to reviewing the changes requested above, and seeing how `sugar` is adopted by the wider community.
