### 2025-05-13 dc7a823
fix SyntaxWarning, see #2
