### 2025-05-13 5600fc9
mention skipped tests and xfails, closes #2
