### 2025-05-13 rnajena/sugar#2 @trichter
Yes, you see a different set of skipped tests, because I was recently able to add dna_features_viewer as a dependency in a version that does not depend on biopython.
