### 2025-05-13 9bbeb8a
add remark about minimal install
