### 2025-05-13 7e465e7
document fname=None, fixes #4
