### 2025-05-13 3b7ad71
mention fname=None in contributing guide
