### 2025-05-13 rnajena/sugar#8 @trichter
Yes, this can be done in a more explicit way.

```py
from sugar import read, read_fts
seqs = read('GCA_040267505.1_UoG_LM41_1.0_genomic.fna')
fts = read_fts('GCA_040267505.1_UoG_LM41_1.0_genomic.gff')
seqs.fts = fts
```
I need to mention this in the tutorial.
