### 2025-05-13 b10fa09
debug aspect/symbol size bug on MacOS
