### 2025-05-13 5efd17f
add test checking font size in seq alignment
