+++ src/sugar/tests/test_imaging_alignment.py
@@ -41,3 +41,33 @@ def test_plot_alignment_examples(outdir):
     seqs2 = seqs[:5, :150].copy()
     seqs2.translate(complete=True).plot_alignment(
         outdir / 'ali3.png', color='flower', figsize=(10,4), symbols=True, aspect=2, alpha=0.5, xticks=False, edgecolors='w')
+
+
+def test_aspect_symbol_size(outdir):
+    pytest.importorskip('matplotlib', reason='require matplotlib module')
+    import matplotlib.pyplot as plt
+    seqs = read()[:, :100]
+
+    for adjustable in ('box', 'datalim'):
+        fig = plt.figure(figsize=(10, 8), dpi=100)
+        ax = fig.add_subplot()
+        ax.set_adjustable(adjustable)
+        ax.set_xlim(-10, 150)
+        ax.set_xlim(-2, 15)
+        seqs.plot_alignment(outdir / f'test_plot_ali_aspect1_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, extent=[0, 10, 0, 10], show_spines=True)
+        assert ax.get_aspect() == 0.04
+        text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
+        assert round(text.get_fontsize(), 1) == 4.6
+        plt.close(fig)
+
+    seqs = seqs[:, :2]
+    for adjustable in ('box', 'datalim'):
+        fig = plt.figure(figsize=(2, 1), dpi=100)
+        ax = fig.add_subplot()
+        ax.set_adjustable(adjustable)
+        ax.set_xlim(-1, 5)
+        seqs.plot_alignment(outdir / f'test_plot_ali_aspect2_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, show_spines=True)
+        assert ax.get_aspect() == 2
+        text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
+        assert round(text.get_fontsize(), 2) == 19.25
+        plt.close(fig)
