### 2025-05-13 rnajena/sugar#8 @widdowquinn
Thanks @trichter! If this will associate features with the corresponding sequence (as there are two replicons in that example) that's a really neat thing and it's worth showing off about.
