### 2025-05-13 9a0c244
use outdir fixture
