### 2025-05-13 rnajena/sugar#6 @trichter
This bug is pretty nasty because I cannot reproduce it on my machine. It seems to be related to axis size, aspect ratio, and automatic symbol size calculation. I have added another test for the symbol size and also added a debug print call. Could you run `sugar test -v --web` and the relevant part of the tutorial for the joss_review2 branch on your machine and paste the output here?

The output on my machine (and in github actions) is:
```
test_imaging_alignment.py adj box width 15.5 dpi 100.0 plot width 100.0 xlim 100.0 n 100 symbol_size 15.5 aspect 2.0
adj box width 15.5 dpi 100.0 plot width 100.0 xlim 100.0 n 100 symbol_size 15.5 aspect 2.0
adj box width 15.5 dpi 100.0 plot width 100.0 xlim 100.0 n 100 symbol_size 15.5 aspect 2.0
adj box width 15.5 dpi 100.0 plot width 100.0 xlim 100.0 n 100 symbol_size 15.5 aspect 2.0
adj box width 15.5 dpi 100.0 plot width 100.0 xlim 100.0 n 100 symbol_size 15.5 aspect None
adj box width 15.5 dpi 100.0 plot width 100.0 xlim 100.0 n 100 symbol_size 15.5 aspect None
adj box width 15.5 dpi 100.0 plot width 50.0 xlim 100.0 n 100 symbol_size 7.75 aspect 0.1111111111111111
.adj box width 10.0 dpi 100.0 plot width 50.0 xlim 50.0 n 50 symbol_size 20.0 aspect 2.0
adj box width 10.0 dpi 100.0 plot width 50.0 xlim 50.0 n 50 symbol_size 20.0 aspect 2.0
.adj box width 7.75 dpi 100 plot width 10.0 xlim 17.0 n 100 symbol_size 4.5588235294117645 aspect 0.04
adj datalim width 7.75 dpi 100 plot width 10.0 xlim 17.0 n 100 symbol_size 4.5588235294117645 aspect 0.04
adj box width 1.155 dpi 100 plot width 2.0 xlim 6.0 n 2 symbol_size 19.25 aspect 2.0
adj datalim width 1.55 dpi 100 plot width 2.0 xlim 8.051948051948052 n 2 symbol_size 19.25 aspect 2.0
.
```

For the first part of the tutorial:
```
In [1]: seqs = read('https://osf.io/download/j2wyv')
In [2]: seqs[:, 70:120].plot_alignment(fname='ali.pdf', color=None, figsize=(10,8), symbols=True, aspect=\
   ...: 2, alpha=0.5)
adj box width 6.451612903225808 dpi 100.0 plot width 50.0 xlim 50.0 n 50 symbol_size 12.903225806451614 aspect 2.0
```

Thank you!
