### 2025-05-14 250e23c
fix genbank location parsing, add test, closes #5
