### 2025-05-14 e0b01af
genbank: allow parsing seqids inside location fields
