### 2025-05-14 c2ecf56
doc that features are not converted to and from BioPython, see #7
