### 2025-05-14 b24e172
update tutorial to use provided example files
