### 2025-05-15 2ca6fd8
fix auto-detection when archiving in special cases
