### 2025-05-15 67d0b60
add tutorial section about fts-seqs association, closes #8
