### 2025-05-15 rnajena/sugar#7 @trichter
For now, I have only documented that features are not preserved. This is definitely something I want to add in a future release.
