### 2025-05-15 c244771
add tutorial command
