### 2025-05-15 e0f8ee4
add test for sugar tutorial command
