### 2025-05-15 dd073e1
add some common fasta file extensions
