+++ src/sugar/_io/fasta.py
@@ -8,7 +8,7 @@ from sugar import BioSeq
 from sugar._io.util import _add_fmt_doc
 
 
-filename_extensions_fasta = ['fasta', 'fa']
+filename_extensions_fasta = ['fasta', 'fa', 'fna', 'faa', 'fas']
 
 def is_fasta(f, **kw):
     content = f.read(50)
