### 2025-05-15 494582b
fix: warning not raised anymore
