### 2025-05-23 7348e1e
doc: minor rename
