### 2025-05-23 7bd7154
move adapter methods to new module
