+++ src/sugar/core/seq.py
@@ -542,7 +542,7 @@ class BioSeq():
     @classmethod
     def frombiopython(cls, obj):
         """
-        Create a `BioSeq` object from a biopython_ `~Bio.SeqRecord.SeqRecord` or `~Bio.Seq.Seq` object.
+        Create a `.BioSeq` object from a biopython_ `~Bio.SeqRecord.SeqRecord` or `~Bio.Seq.Seq` object.
 
         :param obj: The object to convert.
 
@@ -550,30 +550,18 @@ class BioSeq():
             BioPython Features in the ``SeqRecord.features`` attribute are not converted.
             This functionality may be added in a future release of sugar.
         """
-        if hasattr(obj, 'seq'):  # SeqRecord
-            data = str(obj.seq)
-            id_ = obj.id
-        else:  # Seq
-            data = str(obj)
-            id_ = None
-        return cls(data, id=id_)
+        from sugar.core._adapter import biopython2seq
+        return biopython2seq(obj, cls=cls)
 
     @classmethod
     def frombiotite(cls, obj):
         """
-        Create a `BioSeq` object from a biotite_ sequence object.
+        Create a `.BioSeq` object from a biotite_ sequence object.
 
         :param obj: The object to convert.
         """
-        from biotite.sequence import NucleotideSequence, ProteinSequence
-        data = ''.join(obj.symbols)
-        type_ = None
-        if isinstance(obj, NucleotideSequence):
-            type_ = 'nt'
-        elif isinstance(obj, ProteinSequence):
-            type_ = 'aa'
-        return cls(data, type=type_)
-
+        from sugar.core._adapter import biotite2seq
+        return biotite2seq(obj, cls=cls)
 
     def match(self, *args, **kw):
         """
@@ -626,11 +614,10 @@ class BioSeq():
 
         Attached ``BioSeq.fts`` features are not converted.
         """
-        from Bio.Seq import Seq
-        from Bio.SeqRecord import SeqRecord
-        return SeqRecord(Seq(self.data), id=self.id)
+        from sugar.core._adapter import seq2biopython
+        return seq2biopython(self)
 
-    def tobiotite(self, type=None, gap='-', warn=True):
+    def tobiotite(self, **kw):
         """
         Convert BioSeq to biotite_ `~biotite.sequence.NucleotideSequence` or `~biotite.sequence.ProteinSequence` instance
 
@@ -640,20 +627,8 @@ class BioSeq():
         :param str gap: Gap characters that must be removed from the sequence string.
         :param bool warn: Whether to warn if gap characters have been removed, default is True.
         """
-        from biotite.sequence import NucleotideSequence, ProteinSequence
-        data = self.data
-        if gap:
-            for g in gap:
-                if g in data:
-                    if warn:
-                        from warnings import warn
-                        warn(f'Remove gap characters {gap} for the conversion to biotite')
-                    for g in gap:
-                        data = data.replace(g, '')
-                    break
-        type = type or self.type
-        cls = {'nt': NucleotideSequence, 'aa': ProteinSequence}[type]
-        return cls(data)
+        from sugar.core._adapter import seq2biotite
+        return seq2biotite(self, **kw)
 
     def toftsviewer(self, **kw):
         r"""
@@ -1069,7 +1044,7 @@ class BioBasket(collections.UserList):
     @classmethod
     def frombiopython(cls, obj):
         """
-        Create a `BioBasket` object from a list of biopython_ `~Bio.SeqRecord.SeqRecord` or `~Bio.Seq.Seq` objects.
+        Create a `.BioBasket` object from a list of biopython_ `~Bio.SeqRecord.SeqRecord` or `~Bio.Seq.Seq` objects.
 
[…]
