### 2025-05-23 dfd0567
add new test module for adapters
