### 2025-05-26 4765ed4
allow to convert features to/from biopython, closes #7
