+++ src/sugar/tests/test_core_adapter.py
@@ -1,10 +1,12 @@
-# (C) 2024, Tom Eulenfeld, MIT license
+# (C) 2025, Tom Eulenfeld, MIT license
 
+from importlib.resources import files
 import pytest
-from sugar import read
+from sugar import BioBasket, BioSeq, Feature, read, read_fts
+from sugar.tests.util import _clean_fts, _clean_seqs
 
 
-def test_bioseq_to_from_biopython():
+def test_seq2biopython2seq():
     SeqRecord = pytest.importorskip('Bio.SeqRecord', reason='require biopython module')
     seqs = read()
     seq = seqs[0]
@@ -13,9 +15,10 @@ def test_bioseq_to_from_biopython():
     assert isinstance(obj, SeqRecord.SeqRecord)
     assert seq2.id == obj.id == seq.id
     assert str(seq2) == str(obj.seq) == str(seq)
+    assert _clean_seqs([seq])[0] == seq2
 
 
-def test_biobasket_to_from_biopython():
+def test_seqs2biopython2seqs():
     SeqRecord = pytest.importorskip('Bio.SeqRecord', reason='require biopython module')
     seqs = read()
     obj = seqs.tobiopython()
@@ -23,9 +26,30 @@ def test_biobasket_to_from_biopython():
     assert isinstance(obj[0], SeqRecord.SeqRecord)
     assert seqs2[0].id == obj[0].id == seqs[0].id
     assert str(seqs2[0]) == str(obj[0].seq) == str(seqs[0])
+    assert _clean_seqs(seqs) == seqs2
 
 
-def test_biobasket_to_from_biopython_msa():
+def test_biopython2seqs2biopython():
+    SeqIO = pytest.importorskip('Bio.SeqIO', reason='require biopython module')
+    fname = files('sugar.tests.data').joinpath('example.gb')
+    records = list(SeqIO.parse(fname, 'genbank'))
+    seqs = BioBasket.frombiopython(records)
+    assert [len(seq.fts) for seq in seqs] == [len(seq.features) for seq in records]
+    records2 = seqs.tobiopython()
+    for attr in 'seq id name description dbxrefs annotations'.split():
+        assert getattr(records2[0], attr) == getattr(records[0], attr)
+    from Bio.SeqRecord import Seq, SeqRecord
+    record = SeqRecord(Seq('ATG'))
+    seq = BioSeq.frombiopython(record)
+    assert seq.id is None
+    assert '_biopython' not in seq.meta
+    assert seq.data == 'ATG'
+    record2 = seq.tobiopython()
+    for attr in 'seq id name description dbxrefs annotations'.split():
+        assert getattr(record2, attr) == getattr(record, attr)
+
+
+def test_seqs2biopython2seqs_msa():
     SeqRecord = pytest.importorskip('Bio.SeqRecord', reason='require biopython module')
     seqs = read().str.ljust(10_000, '-')
     obj = seqs.tobiopython(msa=True)
@@ -33,6 +57,38 @@ def test_biobasket_to_from_biopython_msa():
     assert isinstance(obj[0], SeqRecord.SeqRecord)
     assert seqs2[0].id == obj[0].id == seqs[0].id
     assert str(seqs2[0]) == str(obj[0].seq) == str(seqs[0])
+    assert _clean_seqs(seqs) == seqs2
+
+
+def test_fts2biopython2fts():
+    SeqFeature = pytest.importorskip('Bio.SeqFeature', reason='require biopython module')
+    from sugar.core._adapter import fts2biopython, biopython2fts
+    fts = read_fts()
+    obj = fts2biopython(fts)
+    fts2 = biopython2fts(obj)
[…]
