### 2025-05-26 2699c6c
fix warning in test: biopython does not close file
