### 2025-05-26 rnajena/sugar#6 @widdowquinn
Hi @trichter - the `matplotlib` version is:

```bash
% conda list | grep matplotlib                                                                                                                            [3:27:23]
matplotlib                3.10.3                   pypi_0    pypi
```

It may be relevant that, on retesting just now with `sugar test`, I see a single test failure in where the fontsize is not returned as expected:

```bash
% pytest src/sugar/tests/test_imaging_alignment.py::test_aspect_symbol_size                                                                               [3:30:36]
========================================================================================== test session starts ==========================================================================================
platform darwin -- Python 3.13.3, pytest-8.3.5, pluggy-1.5.0
rootdir: /Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/tests
configfile: pytest.ini
collected 1 item

src/sugar/tests/test_imaging_alignment.py F                                                                                                                                                       [100%]

=============================================================================================== FAILURES ================================================================================================
________________________________________________________________________________________ test_aspect_symbol_size ________________________________________________________________________________________

outdir = PosixPath('/private/var/folders/9m/72vj0mg57w5_dn3k2qj578500000gn/T/[REDACTED]/pytest-3/test_aspect_symbol_size0')

    def test_aspect_symbol_size(outdir):
        pytest.importorskip('matplotlib', reason='require matplotlib module')
        import matplotlib.pyplot as plt
        seqs = read()[:, :100]

        for adjustable in ('box', 'datalim'):
            fig = plt.figure(figsize=(10, 8), dpi=100)
            ax = fig.add_subplot()
            ax.set_adjustable(adjustable)
            ax.set_xlim(-10, 150)
            ax.set_xlim(-2, 15)
            seqs.plot_alignment(outdir / f'test_plot_ali_aspect1_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, extent=[0, 10, 0, 10], show_spines=True)
            assert ax.get_aspect() == 0.04
            text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
>           assert round(text.get_fontsize(), 1) == 4.6
E           AssertionError: assert 9.1 == 4.6
E            +  where 9.1 = round(9.117647058823529, 1)
E            +    where 9.117647058823529 = get_fontsize()
E            +      where get_fontsize = Text(0.05, 7.5, 'A').get_fontsize

src/sugar/tests/test_imaging_alignment.py:60: AssertionError
======================================================================================== short test summary info ========================================================================================
FAILED src/sugar/tests/test_imaging_alignment.py::test_aspect_symbol_size - AssertionError: assert 9.1 == 4.6
=========================================================================================== 1 failed in 0.32s ===========================================================================================
```
