### 2025-05-26 rnajena/sugar#6 @trichter
Yes, that is related. The test failure and your images with the large font are both incarnations of the same bug. (Tells me, that the test is working as expected.) The same test is passing on githubactions/MacOS with the same matplotlib version.
Do you have any custom matplotlib configuration (matplotlibrc)?
