### 2025-05-26 #8122 @widdowquinn
This all looks very nice, now. I liked the `sugar tutorial` command for moving files to the current location (though I did first look for a download link ;) ). It would be nice to have a brief note in the tutorial like that for `read()` to indicate that `read_fts()` called without arguments also returns example data, but I'm not asking for revision. Everything else is resolved to my satisfaction.

This is a very impressive framework, congratulations! I wish you every success with it!
