### 2025-05-26 rnajena/sugar#6 @widdowquinn
I don't have any custom settings, I'm afraid. This is a new raw installation in its own environment for reviewing. It's a relatively recent rebuild of the machine too and otherwise by far the cleanest environment I have in terms of potential stray config files/settings.
