### 2025-05-27 4f76fdc
gff: be nice on parsing errors, better error msg with line number
