### 2025-05-27 1f3307e
gff: test for parsing error
