+++ src/sugar/tests/test_io_gff.py
@@ -1,5 +1,6 @@
 # (C) 2024, Tom Eulenfeld, MIT license
 
+import pytest
 from sugar import read, read_fts
 from sugar.tests.util import _clean_fts, tempfilename
 
@@ -33,6 +34,15 @@ def test_gff_quote():
     assert fts2[0].seqid == fts[0].seqid
     assert fts2[0].name == fts[0].name
     assert ':' in fts.tofmtstr('gff')
+
+
+def test_gff_parsing_error():
+    with pytest.raises(ValueError, match='Parsing error'):
+        fts = read_fts('!data/io_gff_parsing_error.gff')
+    with pytest.raises(ValueError, match='Parsing error'):
+        fts = read_fts('!data/io_gff_parsing_error.gff', 'gtf')
+    with pytest.raises(ValueError, match='Parsing error'):
+        fts = read_fts('!data/io_gff_parsing_error.gff', 'gff', gff_version='3')
 
 
     # TODO test write and load again
