### 2025-05-27 7fe4bcf
make symbol font size tests more robust, add debug message for failing tests
