+++ src/sugar/tests/test_imaging_alignment.py
@@ -43,31 +43,44 @@ def test_plot_alignment_examples(outdir):
         outdir / 'ali3.png', color='flower', figsize=(10,4), symbols=True, aspect=2, alpha=0.5, xticks=False, edgecolors='w')
 
 
+def _debug_wrong_symbol_size(ax, x, n):
+    # error debug message, see issue #6
+    fig = ax.get_figure()
+    dpiinv = fig.dpi_scale_trans.inverted()
+    bbox = ax.get_window_extent().transformed(dpiinv)
+    symbol_size = bbox.width * fig.dpi * abs((x[-1] - x[0]) / (ax.get_xlim()[1] - ax.get_xlim()[0])) / n
+    msg = ('Wrong symbol size\n'
+           f'{symbol_size=:.2f} = {bbox.width=:} * {fig.dpi=:} * abs({(x[-1] - x[0])} / ({ax.get_xlim()[1]=:.2f} - {ax.get_xlim()[0]=:.2f})) / {n}')
+    return msg
+
+
 def test_aspect_symbol_size(outdir):
     pytest.importorskip('matplotlib', reason='require matplotlib module')
     import matplotlib.pyplot as plt
+    import matplotlib as mpl
     seqs = read()[:, :100]
 
-    for adjustable in ('box', 'datalim'):
-        fig = plt.figure(figsize=(10, 8), dpi=100)
-        ax = fig.add_subplot()
-        ax.set_adjustable(adjustable)
-        ax.set_xlim(-10, 150)
-        ax.set_xlim(-2, 15)
-        seqs.plot_alignment(outdir / f'test_plot_ali_aspect1_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, extent=[0, 10, 0, 10], show_spines=True)
-        assert ax.get_aspect() == 0.04
-        text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
-        assert round(text.get_fontsize(), 1) == 4.6
-        plt.close(fig)
+    with mpl.rc_context({'figure.subplot.left': 0.1, 'figure.subplot.right':  0.9}):
+        for adjustable in ('box', 'datalim'):
+            fig = plt.figure(figsize=(12.5, 8), dpi=100)
+            ax = fig.add_subplot()
+            ax.set_adjustable(adjustable)
+            ax.set_xlim(-5, 15)
+            ax.set_ylim(-2, 12)
+            seqs.plot_alignment(outdir / f'test_plot_ali_aspect1_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, extent=[0, 10, 0, 10], show_spines=True)
+            assert ax.get_aspect() == 0.04
+            text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
+            assert round(text.get_fontsize(), 1) == 5.0, _debug_wrong_symbol_size(ax, x=(0, 10), n=100)
+            plt.close(fig)
 
-    seqs = seqs[:, :2]
-    for adjustable in ('box', 'datalim'):
-        fig = plt.figure(figsize=(2, 1), dpi=100)
-        ax = fig.add_subplot()
-        ax.set_adjustable(adjustable)
-        ax.set_xlim(-1, 5)
-        seqs.plot_alignment(outdir / f'test_plot_ali_aspect2_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, show_spines=True)
-        assert ax.get_aspect() == 2
-        text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
-        assert round(text.get_fontsize(), 2) == 19.25
-        plt.close(fig)
[…]
