### 2025-05-27 e9f16f3
add layout configuration to test
