### 2025-05-27 1db803a
tests: warn when outdir directory is created
