### 2025-05-27 rnajena/sugar#6 @trichter
I still don't have a clue. Could you help me debug again?
If so, please do the following:
Pull the latest changes from master. Run the tests with
```
sugar test --outdir=/tmp/test_sugar
```
`/tmp/test_sugar` should be the absolute path to a temporary directory.
Paste the test error here and upload the image `test_plot_ali_aspect1_box.png` created in the temporary directory.

![Image](https://github.com/user-attachments/assets/0fee6ec5-5370-4a1e-95a6-8697206639f9)
