### 2025-05-27 rnajena/sugar#6 @widdowquinn
Hi @trichter - here's what I get.

```bash
% git logone                                                                                                                 [4:00:01]
1db803abdb59a8f98cacf96ab38295185c312d46 (HEAD -> master, origin/master, origin/HEAD) tests: warn when outdir directory is created
e9f16f3d3248b6a0ac12f1406b0e9868d7993a49 add layout configuration to test
[...]
% sugar test --outdir=2025-05-27_test
==================================================================================== FAILURES =====================================================================================
_____________________________________________________________________________ test_aspect_symbol_size _____________________________________________________________________________

outdir = PosixPath('2025-05-27_test')

    def test_aspect_symbol_size(outdir):
        mpl = pytest.importorskip('matplotlib', reason='require matplotlib module')
        import matplotlib.pyplot as plt
        seqs = read()[:, :100]

        with mpl.rc_context({'figure.autolayout': False,'figure.constrained_layout.use': False,
                             'figure.subplot.left': 0.1, 'figure.subplot.right':  0.9}):
            for adjustable in ('box', 'datalim'):
                fig = plt.figure(figsize=(12.5, 8), dpi=100)
                ax = fig.add_subplot()
                ax.set_adjustable(adjustable)
                ax.set_xlim(-5, 15)
                ax.set_ylim(-2, 12)
                seqs.plot_alignment(outdir / f'test_plot_ali_aspect1_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, extent=[0, 10, 0, 10], show_spines=True)
                assert ax.get_aspect() == 0.04
                text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
>               assert round(text.get_fontsize(), 1) == 5.0, _debug_wrong_symbol_size(ax, x=(0, 10), n=100)
E               AssertionError: Wrong symbol size
E                 symbol_size=10.00 = bbox.width=10.0 * fig.dpi=200.0 * abs(10 / (ax.get_xlim()[1]=15.00 - ax.get_xlim()[0]=-5.00)) / 100
E               assert 10.0 == 5.0
E                +  where 10.0 = round(10.0, 1)
E                +    where 10.0 = get_fontsize()
E                +      where get_fontsize = Text(0.05, 7.5, 'A').get_fontsize

test_imaging_alignment.py:73: AssertionError
================================================================================ warnings summary =================================================================================
test_imaging_alignment.py::test_plot_alignment
  /Users/[REDACTED]/Refereeing/2025-05-12_JOSS/sugar/src/sugar/tests/conftest.py:71: UserWarning: Create missing outdir directory at 2025-05-27_test
    warn(f'Create missing outdir directory at {out}')

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
============================================================================= short test summary info =============================================================================
FAILED test_imaging_alignment.py::test_aspect_symbol_size - AssertionError: Wrong symbol size
========================================================= 1 failed, 146 passed, 10 skipped, 2 xfailed, 1 warning in 3.88s =========================================================
```

and the image produced in `test_plot_ali_aspect1_box.png` is:

![Image](https://github.com/user-attachments/assets/b6502ae0-6afe-4097-ab2a-101d03cdba93)
