### 2025-05-28 f0d16bf
add contains, distance, overlaps, overlaplen, mid methods, add tests, add changelog, fix for mmseqs and co parsing, deprecate LocationTuple.__sub__
