+++ src/sugar/core/fts.py
@@ -127,6 +127,17 @@ class Location():
             self._meta = Meta()
         return self._meta
 
+    @property
+    def range(self):
+        """
+        Get the range of the location
+
+        :returns:
+            tuple ``start, stop`` with start and stop location
+            (zero-based numbering)
+        """
+        return self.start, self.stop
+
     @meta.setter
     def meta(self, v):
         self._meta = None if v is None else Meta(v)
@@ -172,6 +183,99 @@ class Location():
         defect = loc.defect._reverse()
         return Location(start, stop, strand, defect, meta=loc.meta)
 
+    def __lt__(self, other):
+        if isinstance(other, (Location, LocationTuple)):
+            start, stop = self.range
+            start2, stop2 = other.range
+            return start < start2 or (start==start2 and stop < stop2)
+        msg = f"'<' not supported between instances of '{type(self).__name__}' and '{type(other).__name__}'"
+        raise TypeError(msg)
+
+    def __le__(self, other):
+        if isinstance(other, (Location, LocationTuple)):
+            start, stop = self.range
+            start2, stop2 = other.range
+            return start < start2 or (start==start2 and stop <= stop2)
+        msg = f"'<=' not supported between instances of '{type(self).__name__}' and '{type(other).__name__}'"
+        raise TypeError(msg)
+
+    def __gt__(self, other):
+        if isinstance(other, (Location, LocationTuple)):
+            start, stop = self.range
+            start2, stop2 = other.range
+            return start > start2 or (start==start2 and stop > stop2)
+        msg = f"'>' not supported between instances of '{type(self).__name__}' and '{type(other).__name__}'"
+        raise TypeError(msg)
+
+    def __ge__(self, other):
+        if isinstance(other, (Location, LocationTuple)):
+            start, stop = self.range
+            start2, stop2 = other.range
+            return start > start2 or (start==start2 and stop >= stop2)
+        msg = f"'>=' not supported between instances of '{type(self).__name__}' and '{type(other).__name__}'"
+        raise TypeError(msg)
+
+    def mid(self, round=True):
+        """
+        Return the middle position of the location or location tuple
+        """
+        s = sum(self.range)
+        return s // 2 if round else s / 2
+
+    def overlaps(self, other):
+        """
+        Whether the location/locations overlap with the other location/locations
+        """
+        if isinstance(other, (Location, LocationTuple)):
+            lr1 = self.range
+            lr2 = other.range
+            return lr1[0] < lr2[1] and lr1[1] > lr2[0]
+        msg = f"{type(self).__name__}.overlaps() not supported for instances of '{type(other).__name__}'"
+        raise TypeError(msg)
+
+    def overlaplen(self, other):
+        """
+        Return overlap length with the other location or location tuple
+        """
+        if isinstance(other, (Location, LocationTuple)):
+            lr1 = self.range
+            lr2 = other.range
+            return max(0, min(lr1[1], lr2[1]) - max(lr1[0], lr2[0]))
+        msg = f"{type(self).__name__}.overlaplen() not supported for instances of '{type(other).__name__}'"
+        raise TypeError(msg)
+
+    def contains(self, other):
+        """
+        Whether the location range contains the other location range
+        """
+        if isinstance(other, (Location, LocationTuple)):
+            lr1 = self.range
+            lr2 = other.range
+            return lr1[0] <= lr2[0] and lr2[1] <= lr1[1]
+        msg = f"{type(self).__name__}.overlaps() not supported for instances of '{type(other).__name__}'"
+        raise TypeError(msg)
+
+    def distance(self, other, *, pos='inner', sign=False):
+        """
+        Distance to other location or location tuple
+        """
+        assert pos in ('inner', 'middle')
[…]
