### 2025-06-02 4ca9cef
make mid an attribute
