### 2025-06-02 5da48c6
add start and stop attributes to LocationTuple
