### 2025-06-02 4838795
potentially fix fig.dpi symbol size issue
