### 2025-06-02 edb899a
tutorial: mention that read_fts is called without arguments
