### 2025-06-03 rnajena/sugar#6 @widdowquinn
Hmm, I'm still getting this (on my laptop just now, rather than the machine I was previously testing on):

```python
Python 3.12.10 | packaged by conda-forge | (main, Apr 10 2025, 22:18:27) [Clang 18.1.8 ] on darwin
>>> import matplotlib.pyplot as plt
>>> fig = plt.figure(dpi=100)
>>> print(fig.dpi)
200.0
```

and

```python
>>> fig = plt.figure(dpi=100)
>>> ax.set_adjustable('box')
>>> fig.draw_without_rendering()
>>> print(fig.dpi)
200.0
```

and the same error with `sugar test --outdir=/tmp/test_sugar`

```python
[...]
>               assert round(text.get_fontsize(), 1) == 5.0, _debug_wrong_symbol_size(ax, x=(0, 10), n=100)
E               AssertionError: Wrong symbol size
E                 symbol_size=10.00 = bbox.width=10.0 * fig.dpi=200.0 * abs(10 / (ax.get_xlim()[1]=15.00 - ax.get_xlim()[0]=-5.00)) / 100
E               assert 10.0 == 5.0
E                +  where 10.0 = round(10.0, 1)
E                +    where 10.0 = get_fontsize()
E                +      where get_fontsize = Text(0.05, 7.5, 'A').get_fontsize

test_imaging_alignment.py:73: AssertionError
[...]
```

This is on edb899a:

```bash
% git lol
* edb899a (HEAD -> master, origin/master, origin/HEAD) tutorial: mention that read_fts is called without arguments
* 4838795 potentially fix fig.dpi symbol size issue
* 5da48c6 add start and stop attributes to LocationTuple
* 9a807e8 add mid test, doc
[...]
```

Is there anything else you'd like me to try, or other info about my setup that might help?
