### 2025-06-03 rnajena/sugar#6 @trichter
> import matplotlib.pyplot as plt
> fig = plt.figure(dpi=100)
> print(fig.dpi)
> 200.0

I don't think this is the expected behavior of matplotlib. Would you mind opening a ticket with matplotlib? Otherwise, I can do it.
