### 2025-06-03 22b0e82
imaging: add scale_symbol_size parameter for alignment plot to adjust the default symbol size, see #6
