+++ src/sugar/tests/test_imaging_alignment.py
@@ -79,8 +79,8 @@ def test_aspect_symbol_size(outdir):
             ax = fig.add_subplot()
             ax.set_adjustable(adjustable)
             ax.set_xlim(-2, 2)
-            seqs.plot_alignment(outdir / f'test_plot_ali_aspect2_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, show_spines=True)
+            seqs.plot_alignment(outdir / f'test_plot_ali_aspect2_{adjustable}.png', ax=ax, symbols=True, color=None, aspect=2, show_spines=True, scale_symbol_size=1/1.925)
             assert ax.get_aspect() == 2
             text = [child for child in ax.get_children() if hasattr(child, 'get_text') and child.get_text() == 'A'][0]
-            assert round(text.get_fontsize(), 2) == 19.25, _debug_wrong_symbol_size(ax, x=(-0.5, 1.5), n=2)
+            assert round(text.get_fontsize(), 1) == 10.0, _debug_wrong_symbol_size(ax, x=(-0.5, 1.5), n=2*1.925)
             plt.close(fig)
