### 2025-06-03 rnajena/sugar#6 @trichter
For now, I added a new kwarg `scale_symbol_size` to `plot_alignment`, which scales the symbol size. In your case, you could work around the problem by setting `scale_symbol_size=0.5`. This parameter may also be useful for minor tweaks in other cases.
