### 2025-06-03 674f0a1
correct symbol size calculation for non-default dpi values, finally fix #6
