### 2025-06-03 60d8ffe
fix internal pytest error in conjunction with --cov sugar -W error
    INTERNALERROR> pluggy.PluggyTeardownRaisedWarning: A plugin raised an exception during an old-style hookwrapper teardown.
    INTERNALERROR> Plugin: _cov, Hook: pytest_runtestloop
    INTERNALERROR> CoverageWarning: Module sugar was previously imported, but not measured (module-not-measured)
