+++ src/sugar/tests/pytest.ini
@@ -2,7 +2,11 @@
 markers =
     webtest: mark a test as a webtest
 filterwarnings =
+    # all filters are added as markers in conftest.py
     ignore:Found U
     # ignore DeprecationWarnings and PendingDeprecationWarnings triggered by other modules
     ignore::DeprecationWarning:(?!sugar).*
     ignore::PendingDeprecationWarning:(?!sugar).*
+    # ignore CoverageWarning, because pytest is not run from command line
+    # this filter is additionally hard-coded in scripts.py
+    ignore:Module sugar was previously imported
