### 2025-06-03 rnajena/sugar#6 @trichter
I think I found the bug and fixed it. If not please, reopen this issue. Thank you!
