### 2025-06-03 301ab55
print sugar version in tests
