### 2025-06-03 #8122 @trichter
@olexandr-konovalov I finished the author tasks. I released a new version, v0.9.0, which is archived at Zenodo: https://doi.org/10.5281/zenodo.15588305.
