### 2025-06-04 ea7de0d
fix empty fts after remove_overlapping with duplicate fts
