### 2025-06-19 20e86ce
scripts: add sugar rc command to get the reverse complement of a nt sequence
