### 2025-06-23 6a32322
allow continue-on-error for web tests
