### 2025-06-24 4d269af
mention python version requirement
