### 2025-07-01 0740354
select method: warn if attribute is not present
