### 2025-07-01 d46caef
allow to initialize BioBasket object with a single sequence, but discourage this with a warning
