+++ src/sugar/core/cane.py
@@ -380,6 +380,14 @@ def find_orfs(seq, rf='all', start='start', stop='stop', need_start='always', ne
     :returns: Returns a `~ORFList` of all found ORFs.
         You can attach these features to sequences using `.BioSeq.add_fts()` or `.BioBasket.add_fts()`.
         Use the `.BioSeq.fts` and `.BioBasket.fts` properties to overwrite features with the found ORFs.
+
+    ..note::
+        Python's `re.matchall()` is used internally to find start and stop codons.
+        The limitations of this function apply; for example, matches cannot overlap.
+        Care must be taken in special cases.
+        For instance, if ORFs do not need to start with a start codon,
+        do not use the regular expression ``start='...'``;
+        use the ``need_start='never'`` option instead.
     """
     # rf  0, 1, 2, -1, -2, -3, 'fwd', 'bwd', 'all'
     assert need_start in ('never', 'always', 'once')
