### 2025-07-01 ad42367
test for gappy find_orfs, add warning in special untreated case
