### 2025-07-01 d88c9d3
raise specific error msg if user tries to read an empty file
