### 2025-07-02 fbd581a
add EMBL flat file read support
