+++ src/sugar/_io/embl.py
@@ -1,0 +1,190 @@
+# (C) 2023, Tom Eulenfeld, MIT license
+"""
+`EMBL`_ flat file reader for ENA and UniProt
+
+``EMBL`` stands for the European Molecular Biology Laboratory.
+``ENA`` stands for the European Nucleotide Archive.
+``UniProt`` stands for the Universal Protein Database.
+"""
+
+from sugar.core.fts import Defect, Feature, FeatureList, Location
+from sugar.core.seq import Attr, BioSeq, Meta
+from sugar._io.util import _add_fmt_doc
+from sugar._io.genbank import _parse_locs
+
+
+def is_embl(f, **kw):
+    content = f.read(5)
+    return content == 'ID   '
+
+
+is_fts_embl = is_embl
+
+
+_EMBL2GENBANK = {
+    'ID': 'locus',
+    'AC': 'accession',
+    'DT': 'date',
+    'DE': 'definition',
+    'GN': 'gene_name',
+    'KW': 'keywords',
+    'DR': 'dbsource',
+    'CC': 'comment',
+    'OS': 'organism',
+    'OC': 'taxonomy',
+    'OG': 'organelle',
+    'PA': 'parent_accession',
+    'PR': 'project',
+    'R': 'reference',
+    }
+
+
+@_add_fmt_doc('read_fts')
+def read_fts_embl(f, exclude=()):
+    """
+    Read EMBL feature records from file into `.FeatureList`
+
+    :param tuple exclude: Tuple of feature names to exclude,
+        possible options are ``'translation', 'fts'``,
+        sequences are excluded anyway.
+    """
+    fts = FeatureList()
+    for seq in iter_embl(f, exclude=('seq',) + exclude):
+        fts.extend(seq.fts)
+    return fts
+
+
+@_add_fmt_doc('read')
+def iter_embl(f, exclude=(), genbank=True):
+    """
+    Read EMBL records and sequences from file into `.BioBasket`
+
+    :param tuple exclude: Tuple of feature names to exclude,
+        possible options are ``'seq', 'translation', 'fts'``
+        or line keys (e.g. ``'CC'``).
+    :param genbank: By default, use genbank like key names in the ``_embl`` meta data,
+        if set to ``False`` will use EMBL two character keys, except for references which will be
+        saved in the ``_embl.R`` attribute.
+    """
+    # allowed entries in exclude: features, translation, seq
+    meta = Meta()
+    attrs = Attr()
+    fts = []
+    misc = []
+    fttype = None
+    ftmeta = None
+    locs = None
+    val = None
+    key = None
+    parse = 'header'
+    seq = ''
+    for line in f:
+        line = line.rstrip()
+        if fttype is not None and (not line.startswith('FT') or len(line[2:20].strip()) > 0):
+            # create a new feature
+            ft = Feature(type=fttype, locs=_parse_locs(locs),
+                            meta=Meta(_embl=ftmeta))
+            fts.append(ft)
+            fttype = None
+        if line.strip() == '' or line[:2] in exclude or line[:2] in ('XX', 'FH'):
+            continue
+        if line.strip() == '//':
+            # create a new sequence object
+            assert len(misc) == 0
+            assert fttype is None
+            if 'AC' in attrs:
+                meta.id = attrs['AC'].split(';')[0].strip()
+                if 'fts' in meta:
+                    for ft in meta.fts:
+                        ft.meta.seqid = meta.id
+            if genbank:
+                meta._embl = {gbkey: attrs[emblkey] for emblkey, gbkey in _EMBL2GENBANK.items() if emblkey in attrs}
+            else:
+                meta._embl = attrs
+            meta.fts = FeatureList(fts)
+            if 'translation' in exclude and 'fts' in meta:
+                for feature in meta.fts:
+                    try:
[…]
