+++ docs/src/sugar._io.embl.rst
@@ -1,0 +1,4 @@
+sugar.\_io.embl module
+======================
+
+.. automodule:: sugar._io.embl

[…]
