### 2025-07-08 9236206
joss: implement editor suggestions
