### 2025-07-09 28c86cb
add read and write support for Clustal format
