+++ pyproject.toml
@@ -41,6 +41,7 @@ sugar = "sugar.scripts:cli"
 fasta = "sugar._io.fasta"
 genbank = "sugar._io.genbank"
 embl = "sugar._io.embl"
+clustal = "sugar._io.clustal"
 stockholm = "sugar._io.stockholm"
 gff = "sugar._io.gff"
 sjson = "sugar._io.sjson"
