### 2025-07-09 c013149
add clustal module to doc
