+++ docs/src/sugar._io.clustal.rst
@@ -1,0 +1,4 @@
+sugar.\_io.clustal module
+=========================
+
+.. automodule:: sugar._io.clustal
