### 2025-07-09 8334cae
add clustal to detection order list
