+++ src/sugar/_io/util.py
@@ -8,7 +8,7 @@ import shutil
 # Lists in this dictionary define the detection order when reading
 # a file without specifying the format.
 # More common and easier to detect formats should be earlier in the list.
-FMTS = {'seqs': ['fasta', 'genbank', 'embl', 'stockholm', 'gff', 'sjson'],
+FMTS = {'seqs': ['fasta', 'genbank', 'embl', 'clustal', 'stockholm', 'gff', 'sjson'],
         'fts': ['gff', 'gtf', 'genbank', 'embl', 'infernal', 'mmseqs', 'blast', 'tsv', 'csv', 'sjson']}
 
 
