### 2025-07-09 #8122 @olexandr-konovalov
@danielskatz but I am pedantic! I did all from "Editor Tasks Prior to Acceptance" and the error only appeared after `@editorialbot recommend-accept` command.
