### 2025-07-09 #8122 @xuanxu
Somehow the use of the `{.python}` modifiers [here](https://github.com/rnajena/sugar/blob/master/docs/joss/paper.md?plain=1#L40-L44) is creating invalid JATS metadata.
