### 2025-07-09 73099b5
update logo script
