### 2025-07-10 7d60524
add dark logo, doc fixes
