+++ .github/workflows/logo.yml
@@ -27,7 +27,7 @@ jobs:
         run: |
           python create_sugar_logo.py sugar_logo.png
           python create_sugar_logo.py sugar_logo_transparent.png --transparent
-          python create_sugar_logo.py sugar_logo_dark.png --transparent --color1 0.9 --color2 0.5
+          python create_sugar_logo.py sugar_logo_dark.png --transparent --color1 0.8 --color2 0.5
       - name: commit logo
         run: |
           git config --global user.name 'trichter'
