### 2025-07-10 #8122 @trichter
@xuanxu That was the Markdown hack I used to create a two-column table with code blocks and syntax highlighting. Do you know of another way to achieve the desired result without creating invalid JATS?
