### 2025-07-11 #8122 @trichter
@xuanxu @olexandr-konovalov @danielskatz If there is no other solution around, I can simply remove the Python modifier. Unfortunately, that would mean loosing the syntax highlighting.
