### 2025-07-12 #8122 @trichter
Yes, I would prefer to keep the code comparison in a two-column table. Another option would be to create it with LaTeX and include the code as a figure.
