### 2025-07-17 daa8821
joss: remove python modifier in code comparison table
