### 2025-07-17 #8122 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
